#600 | refactor(masonry): separate out new Brick type definitions
