+++ modules/masonry/src/components/Brick/BrickFixed.tsx
@@ -6,7 +6,7 @@ import type {
   Size,
   StatementBrickViewProps,
   ValueBrickViewProps,
-} from '@/@types/brick';
+} from '@/@types/brick.types';
 
 import { cn } from '@/lib/utils';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/ui/select';
