+++ modules/masonry/src/components/Brick/BrickInput.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useLayoutEffect, useMemo, useRef, useState } from 'react';
 
-import type { Bounds, Size, ValueBrickViewProps, WidgetInput } from '@/@types/brick';
+import type { Bounds, Size, ValueBrickViewProps, WidgetInput } from '@/@types/brick.types';
 
 import { SCALE_LEVEL_CONFIG } from '@/utils/constants';
 import { BrickOutlineGenerator } from '@/utils/path';
